Added new branch for Ad Ratio Slider
Some other changes to form also.  Removed URL Parameters.  NOTE: Eli, on line 707 of content.js, I had to duplicate a lot of code (from S5 to S51).  I think it should be possible to create arrays with the Slider Likert values, and then make a function for the slider rendering - we'll want many of these.  I'll leave it to you, and make an issue for it.